+++ medical-service/medical-user/src/main/java/com/whz/mapper/RoleMenuMapper.java
@@ -19,9 +19,9 @@ import java.util.List;
 public interface RoleMenuMapper extends BaseMapper<RoleMenu> {
 
     @Select("select menu_id from sys_role_menu where role_id = #{roleId}")
-    List<Integer> selectByRoleId(@Param("roleId")Integer roleId);
+    List<Long> selectByRoleId(@Param("roleId")Long roleId);
 
     @Delete("delete from sys_role_menu where role_id = #{roleId}")
-    int deleteByRoleId(@Param("roleId") Integer roleId);
+    int deleteByRoleId(@Param("roleId") Long roleId);
 
 }
